+++ tsconfig.json
@@ -8,6 +8,6 @@
     "jsx": "react-jsx",
     "outDir": "dist"
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "src/types"],
   "exclude": ["node_modules"]
 }
